+++ renderer/components/dashboard-projects/DashboardProjects.tsx
@@ -1,4 +1,4 @@
-import NextLink from '../../shared/components/responsive-drawer/navigation/nextjs-link/NextLink';
+import NextLink from '../../modules/shared/components/responsive-drawer/navigation/nextjs-link/NextLink';
 import Grid from '@material-ui/core/Grid';
 import Card from '@material-ui/core/Card';
 import CardMedia from '@material-ui/core/CardMedia';
