+++ renderer/modules/home/components/background/background.styles.ts
