+++ renderer/modules/home/components/view-dashboard-projects-detail/ViewDashboardProjectsDetail.styles.ts
