+++ renderer/modules/home/components/view-dashboard-projects-detail/ViewDashboardProjectsDetail.tsx
