+++ renderer/modules/home/components/welcome-snippet/welcome-snippet.tsx
