+++ renderer/modules/shared/components/accept-button/accept-button.styles.ts
