+++ renderer/modules/shared/components/responsive-drawer/ResponsiveDrawer.tsx
@@ -7,10 +7,10 @@ import MenuIcon from '@material-ui/icons/Menu';
 import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Tooltip from '@material-ui/core/Tooltip';
-import AlertIcon from '../../../components/notifications/alert-icon/AlertIcon';
+import AlertIcon from '../../../../components/notifications/alert-icon/AlertIcon';
 import Navigation from './navigation/Navigation';
-import NotificationsDialog from '../../../components/notifications/Notifications';
-import { NotificationsProvider } from '../../../components/notifications/redux/NotificationsContext';
+import NotificationsDialog from '../../../../components/notifications/Notifications';
+import { NotificationsProvider } from '../../../../components/notifications/redux/NotificationsContext';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import CardMedia from '@material-ui/core/CardMedia';
