+++ renderer/pages/installations.tsx
@@ -1,6 +1,6 @@
 import { Component } from 'react';
-import ResponsiveDrawer from '../shared/components/responsive-drawer/ResponsiveDrawer';
-import SpaceAround from '../shared/hoc/SpaceAround';
+import ResponsiveDrawer from '../modules/shared/components/responsive-drawer/ResponsiveDrawer';
+import SpaceAround from '../modules/shared/hoc/SpaceAround';
 import Repository from '../services/github/models/repository.model';
 import Installations from '../components/Installations/Installations.contoller';
